Implements get/set for OutTag value

+++ ScadaSystem/ScadaModels/OutTag.cs
@@ -11,6 +11,6 @@ namespace ScadaModels
     public class OutTag : Tag
     {
         [DataMember]
-        public long InitialValue { get; set; }
+        public double InitialValue { get; set; }
     }
 }
